+++ ablclient/client.go
@@ -4,7 +4,6 @@ import (
 	"context"
 
 	pb "github.com/airbloc/airbloc-go/proto/rpc/v1/userdelegate"
-	"github.com/airbloc/airbloc-go/shared/account"
 	"github.com/airbloc/airbloc-go/shared/key"
 	"github.com/airbloc/airbloc-go/shared/types"
 	ethCommon "github.com/ethereum/go-ethereum/common"
@@ -27,7 +26,7 @@ func NewClient(conn *grpc.ClientConn) *Client {
 	}
 }
 
-func (client *Client) Create(walletAddress ethCommon.Address, password string) (*account.Session, error) {
+func (client *Client) Create(walletAddress ethCommon.Address, password string) (*Session, error) {
 	identity := crypto.Keccak256Hash(walletAddress.Bytes())
 	priv := key.DeriveFromPassword(identity, password)
 
@@ -49,14 +48,14 @@ func (client *Client) Create(walletAddress ethCommon.Address, password string) (
 	if err != nil {
 		return nil, errors.Wrapf(err, "invalid ID returned from the server: %s", response.GetAccountId())
 	}
-	return &account.Session{
+	return &Session{session{
 		AccountId:     accountId,
 		WalletAddress: walletAddress,
 		Key:           priv,
-	}, nil
+	}}, nil
 }
 
-func (client *Client) LogIn(identity string, password string) (*account.Session, error) {
+func (client *Client) LogIn(identity string, password string) (*Session, error) {
 	request := &pb.AccountGetByIdentityRequest{
 		Identity: identity,
 	}
@@ -68,7 +67,7 @@ func (client *Client) LogIn(identity string, password string) (*account.Session,
 	if err != nil {
 		return nil, errors.Wrapf(err, "invalid ID returned from the server: %s", response.GetAccountId())
 	}
-	session := account.NewSession(accountId, ethCommon.HexToAddress(response.GetOwnerAddress()), password)
+	session := NewSession(accountId, ethCommon.HexToAddress(response.GetOwnerAddress()), password)
 
 	// generate test signature
 	identityHash := crypto.Keccak256Hash([]byte(identity))
